requirements.txt
Django==3.2.13
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
django-environ==0.8.1
mysqlclient==2.1.1
requests==2.28.1
python-dateutil==2.8.2
psycopg2-binary==2.9.5
APScheduler==3.9.1
django-allauth==0.51.0
httplib2==0.21.0